+++ lib/render.js
@@ -1,8 +1,31 @@
 'use strict';
 var path = require('path'),
-  Promise = require('bluebird'),
+  _ = require('lodash'),
+  deasync = require('deasync'),
   cons = require('consolidate'),
-  embed = require('./embed');
+  embedTemplate, render;
+
+/**
+ * Tries to embed another template into this template
+ * @param {string} tpl
+ * @param {{}} [data]
+ * @param {{}} [extraData]
+ */
+embedTemplate = function(tpl, data, extraData) {
+  if (!_.isString(tpl)) {
+    return '';
+  }
+
+  data = data || {};
+
+  // Add extra data.
+  if (_.isObject(extraData)) {
+    data = _.cloneDeep(data); // cloneDeep is necessary for embeds within for loops that have extraData.
+    _.defaults(data, extraData);
+  }
+
+  return render(tpl, data);
+};
 
 /**
  * render a layout or component with some data
@@ -11,27 +34,29 @@ var path = require('path'),
  * @return {Promise}
  */
 function render(tpl, data) {
-  var d = Promise.defer();
+  // grab the template file and find out what engine it uses (from the extension)
+  var engine = path.extname(tpl).substring(1), // get rid of the .
+    syncRender = deasync(cons[engine]);
 
   data = data || {};
 
   if (!data.embed) {
     // add embed function to the locals passed into the template
-    data.embed = embed;
+    data.embed = embedTemplate;
   }
 
-  // grab the template file and find out what engine it uses (from the extension)
-  var engine = path.extname(tpl).substring(1); // get rid of the .
-
-  cons[engine](tpl, data, function (err, html) {
-    if (err) {
-      d.reject(err);
-    } else {
-      d.resolve(html);
-    }
-  });
-
-  return d.promise;
+  try {
+    return syncRender(tpl, data);
+  } catch (e) {
+    // this will throw errors if your template is messed up!
+    console.log(e.message, e.stack);
+  }
 }
 
 module.exports = render;
+module.exports.embedTemplate = embedTemplate; // for testing
+
+// DI
+module.exports.setRenderer = function (renderer) {
+  render = renderer;
+};
