+++ package.json
@@ -17,8 +17,8 @@
   },
   "homepage": "https://github.com/nymag/multiplex-templates",
   "dependencies": {
-    "bluebird": "^2.9.21",
     "consolidate": "^0.12.1",
+    "deasync": "0.0.10",
     "glob": "^5.0.3",
     "jade": "^1.9.2",
     "lodash": "^3.5.0",
